noop split templates.go -> templates{,_cluster,_dev,_rp}.go
